move identityX vue components to monorail
