+++ packages/theme-monorail/components/identity-x/comment-stream.marko
@@ -12,7 +12,7 @@ $ const content = getAsObject(input, "content");
     load-more-button-class="btn"
   />
   <marko-web-browser-component
-    name="GlobalCommentToggleButton"
+    name="ThemeCommentToggleButton"
     props={
       identifier: `${content.id}`,
       isExpanded: true,
